requirements.txt
streamlit>=1.32.0
pdf2image>=1.17.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
Pillow>=10.2.0
easyocr>=1.7.1
pyzbar>=0.1.9
pandas>=2.2.0
matplotlib>=3.8.0
scipy>=1.12.0
scikit-image>=0.22.0
